Add generic `Menu` component
